chore: cleanup review step2 and extract calculation table to component
